+++ R/utils.R
@@ -17,13 +17,15 @@ check_factor <- function(x, arg = caller_arg(x), call = caller_env()) {
   } else if (is.factor(x)) {
     x
   } else {
-    cli::cli_abort("{.arg {arg}} must be a factor (or character vector).", call = call)
+    not <- obj_desc(x)
+    cli::cli_abort("{.arg {arg}} must be a factor or character vector, not {not}", call = call)
   }
 }
 
 check_factor_list <- function(x, arg = caller_arg(x), call = caller_env()) {
   if (!is.list(x)) {
-    cli::cli_abort("{.arg {arg}} must be a list", call = call)
+    not <- obj_desc(x)
+    cli::cli_abort("{.arg {arg}} must be a list, not {not}", call = call)
   }
 
   is_factor <- vapply(x, is.factor, logical(1))
@@ -32,4 +34,29 @@ check_factor_list <- function(x, arg = caller_arg(x), call = caller_env()) {
   }
 
   x
+}
+
+obj_desc <- function(x) {
+  if (isS4(x)) {
+    paste0("an S4 object with class ", class(x)[[1]])
+  } else if (is.object(x)) {
+    if (is.data.frame(x)) {
+      "a data frame"
+    } else if (is.factor(x)) {
+      "a factor"
+    } else {
+      paste0("an S3 object with class ", paste(class(x), collapse = "/"))
+    }
+  } else {
+    switch(typeof(x),
+      "NULL" = "a NULL",
+      character = "a character vector",
+      integer = "an integer vector",
+      logical = "a logical vector",
+      double = "a numeric vector",
+      list = "a list",
+      closure = "a function",
+      paste0("a base object of type", typeof(x))
+    )
+  }
 }