+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * `first2()` and `last2()` now ignore missing values in both `x` and `y` (#303).
 
+* Error messages are more informative.
+
 # forcats 0.5.1
 
 * Re-license as MIT (#277).