+++ NEWS.md
@@ -3,6 +3,10 @@
 * New `fct()` which works like `factor()` but errors if values of `x`
   are not included in the levels specification (#299)
 
+* `first2()` and `last2()` now ignore missing values in both `x` and `y` (#303).
+
+* Error messages are more informative.
+
 # forcats 0.5.1
 
 * Re-license as MIT (#277).